Remake of the function selection system
Remake of the function/editing-tool selection system,
the new one is called Toolbox!
Use f-key to move between map-editor-functions, in the toolbox.

I did this so that there will be space to select between more functions,
the keyboard keys were running out :)

+++ app/src/main/java/worldsim/app/Camera.java
@@ -10,8 +10,8 @@ public class Camera {
 	int px = 0;
 	int py = 0;
 	int pz = 0;
-	int cursorx = 10;
-	int cursory = 10;
+	double cursorx = 10;
+	double cursory = 10;
 
 
 	void move(int dx, int dy){
@@ -23,7 +23,7 @@ public class Camera {
 		pz += dz;
 	}
 
-	void moveCursor(int dcursorx, int dcursory) {
+	void moveCursor(double dcursorx, double dcursory) {
 		cursorx += dcursorx;
 		cursory += dcursory;
 	}
